feat: Add styling and script to drop-down menu
The content of the drop-down menu will be add/remove based on item does/ does not display on the main header.
The menu of the link is hidden default now and can be toggle on/of.
Change the background color of breadcrumbs to make it fit with the drop-down menu of header.